Move commands back into cmd packate
When the commands were in the cmd/oscli package rather than in the
cmd/oscli/cmd package, the goreleaser build would cause the root command
to not print out the usage when there were no arguments give. This was
pretty annoying and I couldn't work out exactly why this was happening.
I found that putting the commands back into the cmd/oscli/cmd package,
it make the usage print as expected when no arguments are provided.