master: fix remote config
Prior to this commit, when attempting to configure the remote host or
remote port for sending messages to, it would sometimes not be set,
depending on which subcommand was used.

By putting in some bad practise, of a global variable, this issue has
been fixed.

This should be refactored out later when all package globals are
refactored out.